Bug Fixes
I went through and fixed the combat system a lot. now the AI works and is animated. It should run and is fine.
Sword trails work but are pretty cruddy looking
UserCharacter works well
the Movable system now works on both User and AI (i forgot to initialize the variables in AI version, which caused the enemies not to be animated.)